I have corrected the package declaration in `vm/systemSmartContracts/clob/clobSC.go` from `systemSmartContracts` to `clob`. The file was already in the correct directory, so no move was necessary. I am now ready to move to the next step.
I have implemented the `clobExecutor.go` file and updated the related files, `clobSC.go` and `clob.go`, to use the `Decimal` type and handle order matching. I believe I have completed the implementation of the `clobExecutor.go` file as per the plan. I will now mark this step as complete.

I have confirmed that `clobBuiltin.go` does not exist, and I have removed the old `clobExecutor_test.go` file. This step is now complete.
